CONN-1279c Updated package structure.
