+++ Product/Production/Adapters/CORE_X12DocSubmission_a0/src/main/java/gov/hhs/fha/nhinc/corex12/docsubmission/realtime/adapter/AdapterCORE_X12DSRealTimeImpl.java
@@ -24,25 +24,47 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package gov.hhs.fha.nhinc.corex12.docsubmission.realtime._10.adapter;
+package gov.hhs.fha.nhinc.corex12.docsubmission.realtime.adapter;
 
+import gov.hhs.fha.nhinc.common.nhinccommon.AssertionType;
 import gov.hhs.fha.nhinc.common.nhinccommonadapter.AdapterCOREEnvelopeRealTimeRequestType;
 import gov.hhs.fha.nhinc.common.nhinccommonadapter.AdapterCOREEnvelopeRealTimeResponseType;
-import javax.annotation.Resource;
+import gov.hhs.fha.nhinc.common.nhinccommonadapter.AdapterCOREEnvelopeRealTimeSecuredRequestType;
+import gov.hhs.fha.nhinc.corex12.docsubmission.realtime.adapter.AdapterCORE_X12DSRealTimeOrchImpl;
+
+import gov.hhs.fha.nhinc.messaging.server.BaseService;
 import javax.xml.ws.WebServiceContext;
 
 /**
  *
  * @author sadusumilli
  */
-public class AdapterCORE_X12DSRealTimeUnsecured implements gov.hhs.fha.nhinc.adaptercore.AdapterCORETransactionPortType {
+public class AdapterCORE_X12DSRealTimeImpl extends BaseService {
 
-    @Resource
-    private WebServiceContext context;
+    /**
+     *
+     * @param adapterRequest
+     * @param context
+     * @return
+     */
+    public AdapterCOREEnvelopeRealTimeResponseType realTimeTransaction(AdapterCOREEnvelopeRealTimeRequestType adapterRequest, WebServiceContext context) {
+        AdapterCOREEnvelopeRealTimeResponseType oResponse = new AdapterCOREEnvelopeRealTimeResponseType();
+        AssertionType oAssertion = getAssertion(context, adapterRequest.getAssertion());
+        oResponse.setCOREEnvelopeRealTimeResponse(new AdapterCORE_X12DSRealTimeOrchImpl().realTimeRequest(adapterRequest.getCOREEnvelopeRealTimeRequest(), oAssertion));
+        return oResponse;
+    }
 
-    @Override
-    public AdapterCOREEnvelopeRealTimeResponseType realTimeTransaction(AdapterCOREEnvelopeRealTimeRequestType adapterCOREEnvelopeRealTimeRequestType) {
-        return new AdapterCORE_X12DSRealTimeImpl().realTimeTransaction(adapterCOREEnvelopeRealTimeRequestType, context);
+    /**
+     *
+     * @param adapterRequest
+     * @param context
+     * @return
+     */
+    public AdapterCOREEnvelopeRealTimeResponseType realTimeTransactionSecured(AdapterCOREEnvelopeRealTimeSecuredRequestType adapterRequest, WebServiceContext context) {
+        AdapterCOREEnvelopeRealTimeResponseType oResponse = new AdapterCOREEnvelopeRealTimeResponseType();
+        AssertionType oAssertion = getAssertion(context, null);
+        oResponse.setCOREEnvelopeRealTimeResponse(new AdapterCORE_X12DSRealTimeOrchImpl().realTimeRequest(adapterRequest.getCOREEnvelopeRealTimeRequest(), oAssertion));
+        return oResponse;
     }
 
 }
